+++ p2p/module.go
@@ -55,8 +55,8 @@ type p2pModule struct {
 	pstoreProvider        providers.PeerstoreProvider
 
 	// Assigned during `#Start()`. TLDR; `host` listens on instantiation.
-	// and `network` depends on `host`.
-	network typesP2P.Network
+	// and `router` depends on `host`.
+	router typesP2P.Router
 	// host represents a libp2p network node, it encapsulates a libp2p peerstore
 	// & connection manager. `libp2p.New` configures and starts listening
 	// according to options. Assigned via `#Start()` (starts on instantiation).
@@ -82,7 +82,7 @@ func WithHostOption(host libp2pHost.Host) modules.ModuleOption {
 }
 
 func (m *p2pModule) Create(bus modules.Bus, options ...modules.ModuleOption) (modules.Module, error) {
-	logger.Global.Debug().Msg("Creating libp2p-backed network module")
+	logger.Global.Debug().Msg("Creating P2P module")
 	*m = p2pModule{
 		cfg:    bus.GetRuntimeMgr().GetConfig().P2P,
 		logger: logger.Global.CreateLoggerForModule(modules.P2PModuleName),
@@ -143,7 +143,7 @@ func (m *p2pModule) GetModuleName() string {
 }
 
 // Start instantiates and assigns `m.host`, unless one already exists, and
-// `m.network` (which depends on `m.host` as a required config field).
+// `m.router` (which depends on `m.host` as a required config field).
 func (m *p2pModule) Start() (err error) {
 	m.GetBus().
 		GetTelemetryModule().
@@ -168,8 +168,8 @@ func (m *p2pModule) Start() (err error) {
 		}
 	}
 
-	if err := m.setupNetwork(); err != nil {
-		return fmt.Errorf("setting up network: %w", err)
+	if err := m.setupRouter(); err != nil {
+		return fmt.Errorf("setting up router: %w", err)
 	}
 
 	// Don't handle incoming streams in client debug mode.
@@ -203,7 +203,7 @@ func (m *p2pModule) Broadcast(msg *anypb.Any) error {
 	}
 	m.logger.Info().Msg("broadcasting message to network")
 
-	return m.network.NetworkBroadcast(data)
+	return m.router.Broadcast(data)
 }
 
 func (m *p2pModule) Send(addr cryptoPocket.Address, msg *anypb.Any) error {
@@ -216,7 +216,7 @@ func (m *p2pModule) Send(addr cryptoPocket.Address, msg *anypb.Any) error {
 		return err
 	}
 
-	return m.network.NetworkSend(data, addr)
+	return m.router.Send(data, addr)
 }
 
 // TECHDEBT(#348): Define what the node identity is throughout the codebase
@@ -282,15 +282,17 @@ func (m *p2pModule) setupCurrentHeightProvider() error {
 	return nil
 }
 
-// setupNetwork instantiates the configured network implementation.
-func (m *p2pModule) setupNetwork() (err error) {
-	m.network, err = raintree.NewRainTreeNetwork(
+// setupRouter instantiates the configured router implementation.
+func (m *p2pModule) setupRouter() (err error) {
+	m.router, err = raintree.NewRainTreeRouter(
 		m.GetBus(),
-		raintree.RainTreeConfig{
-			Host:                  m.host,
+		&raintree.RainTreeConfig{
 			Addr:                  m.address,
-			PeerstoreProvider:     m.pstoreProvider,
 			CurrentHeightProvider: m.currentHeightProvider,
+			Host:                  m.host,
+			Hostname:              m.cfg.Hostname,
+			MaxMempoolCount:       m.cfg.MaxMempoolCount,
+			PeerstoreProvider:     m.pstoreProvider,
 		},
 	)
 	return err
@@ -356,10 +358,10 @@ func (m *p2pModule) handleStream(stream libp2pNetwork.Stream) {
 		return
 	}
 
-	if err := m.network.AddPeer(peer); err != nil {
+	if err := m.router.AddPeer(peer); err != nil {
 		m.logger.Error().Err(err).
 			Str("address", peer.GetAddress().String()).
-			Msg("adding remote peer to network")
+			Msg("adding remote peer to router")
 	}
 
 	go m.readStream(stream)
@@ -414,9 +416,9 @@ func (m *p2pModule) readStream(stream libp2pNetwork.Stream) {
 }
 
 // handleNetworkData passes a network message to the configured
-// `Network`implementation for routing.
+// `Router`implementation for routing.
 func (m *p2pModule) handleNetworkData(data []byte) error {
-	appMsgData, err := m.network.HandleNetworkData(data)
+	appMsgData, err := m.router.HandleNetworkData(data)
 	if err != nil {
 		return err
 	}