+++ consensus/doc/CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.0.49] - 2023-04-19
 
-- Add state sync channels `blocksReceived` and `metadataReceived`
-- 
+- Added state sync channels `blocksReceived` and `metadataReceived`, implemented `blockApplicationLoop()`, state sync functions `Start()` and `Stop()`
+- Implemented `WaitForNetworkFSMEvents()` function in test utils
 
 ## [0.0.0.48] - 2023-04-17