+++ consensus/doc/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.49] - 2023-04-27
+
+- Updated comments following P2P refactor
+
 ## [0.0.0.48] - 2023-04-17
 
 - Debug logging improvements